+++ src/main/kotlin/com/vadeen/dns/message/Header.kt
@@ -4,7 +4,7 @@ import com.vadeen.dns.constants.OperationCode
 import com.vadeen.dns.constants.ResponseCode
 
 data class Header(
-    val id: Short,
+    val id: Int,
     val isResponse: Boolean,
     val operationCode: OperationCode,
     val authoritativeAnswer: Boolean,
@@ -12,8 +12,8 @@ data class Header(
     val recursionDesired: Boolean,
     val recursionAvailable: Boolean,
     val responseCode: ResponseCode,
-    val questionRecords: Short,
-    val answerRecords: Short,
-    val authorityRecords: Short,
-    val additionalRecords: Short
+    val questionRecords: Int,
+    val answerRecords: Int,
+    val authorityRecords: Int,
+    val additionalRecords: Int
 )
